API - removed default method arguments

+++ API/Infrastructure/Persistence/Repositories/BasketRepository.cs
@@ -2,6 +2,7 @@
 
 using Domain.Entities.Basket;
 using Domain.Entities.Product;
+using Domain.Entities.User;
 using Domain.Interfaces.Repository;
 
 using Microsoft.EntityFrameworkCore;
@@ -9,8 +10,10 @@ using Microsoft.EntityFrameworkCore;
 public class BasketRepository(StoreContext storeContext) : IBasketRepository
 {
 
-    public Task<Basket?> GetBasketAsync(Guid userId, bool isTracked = false)
+    public Task<Basket?> GetBasketAsync(Guid userId, bool isTracked)
     {
+        if(userId == Guid.Empty) throw new ArgumentNullException(nameof(userId));
+
         var query = isTracked
             ? storeContext.Baskets.AsTracking()
             : storeContext.Baskets.AsNoTracking();
@@ -24,6 +27,7 @@ public class BasketRepository(StoreContext storeContext) : IBasketRepository
 
     public async Task<Basket> AddBasketAsync(Basket basket)
     {
+        ArgumentNullException.ThrowIfNull(basket);
         await storeContext.Baskets.AddAsync(basket);
         await storeContext.SaveChangesAsync();
         return basket;
@@ -31,12 +35,16 @@ public class BasketRepository(StoreContext storeContext) : IBasketRepository
 
     public async Task<bool> UpdateBasketAsync(Basket existingBasket)
     {
+        ArgumentNullException.ThrowIfNull(existingBasket);
         storeContext.Baskets.Update(existingBasket);
         return await storeContext.SaveChangesAsync() != 0;
     }
 
     public async Task<bool> AddItemAsync(Basket existingBasket, Product existingProduct, int quantity)
     {
+        ArgumentNullException.ThrowIfNull(existingBasket);
+        ArgumentNullException.ThrowIfNull(existingProduct);
+
         var existingItem = existingBasket.BasketItems
             .FirstOrDefault(item => item.ProductId == existingProduct.Id);
 
@@ -58,18 +66,21 @@ public class BasketRepository(StoreContext storeContext) : IBasketRepository
 
     public async Task<bool> DeleteBasketAsync(Guid id)
     {
+        if(id == Guid.Empty) throw new ArgumentNullException(nameof(id));
         return await storeContext.Baskets.Where(b => b.Id == id).ExecuteDeleteAsync() != 0;
     }
 
     public async Task<bool> RemoveItemAsync(Basket existingBasket, long productId, int quantity)
     {
+        ArgumentNullException.ThrowIfNull(existingBasket);
+
         var item = existingBasket.BasketItems
             .FirstOrDefault(item => item.ProductId == productId);
         if(item == null) return false;
 
         item.Quantity -= quantity;
         if(item.Quantity <= 0) existingBasket.BasketItems.Remove(item);
-        //TODO: checks for update
+
         return await storeContext.SaveChangesAsync() > 0;
     }
